requirements.txt
aiogram>=3.15.0
aiosqlite>=0.20.0
pandas>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
